// frontend/GrandSmartChef/src/app/models/Enums/ToastType.ts
export enum ToastType {
  SUCCESS = 'success',
  ERROR = 'error',
  WARNING = 'warning',
  INFO = 'info',


  RECIPE_RATED = 'recipe_rated',
  RECIPE_SAVED = 'recipe_saved',
  RECIPE_REMOVED = 'recipe_removed',

  LOGIN_SUCCESS = 'login_success',
  LOGIN_ERROR = 'login_error',

  PROFILE_UPDATED = 'profile_updated',
  PREFERENCES_UPDATED = 'preferences_updated'
}